crypto_price_monitor/requirements.txt
requests
fastapi
sqlalchemy
python-dotenv
schedule
cachetools
sqlalchemy
psycopg2
apscheduler